requirements.txt
# TikTok Monitor — Python 依赖
# 生成时间：2026-03-26

# Web 框架
flask==3.0.3
flask-socketio==5.3.6
flask-login==0.6.3

# TikTok 直播
TikTokLive==6.3.11

# HTTP 客户端（支持 SOCKS5 代理）
httpx[socks]==0.27.0
requests==2.31.0

# 语音识别（Whisper，GPU 版本需要配合 CUDA）
openai-whisper==20231117

# PyTorch（GPU版，在 Docker 内通过 CUDA 加速 Whisper）
# 注意：在 Linux + NVIDIA GPU 环境下，需替换为 CUDA 版本
# torch==2.2.0+cu121  torchvision==0.17.0+cu121
# 安装命令：pip install torch torchvision --index-url https://download.pytorch.org/whl/cu121
torch>=2.0.0

# Excel 报告
openpyxl==3.1.2

# 语言检测
langdetect==1.0.9

# 事件循环（SocketIO 需要）
eventlet==0.35.2

# 工具库
python-dotenv==1.0.1
